format to new worker format
