+++ apps/event-worker/src/index.ts
@@ -1,23 +1,13 @@
 import { logger } from "@ctrlplane/logger";
 
 import { register } from "./instrumentation.js";
-import { createDispatchExecutionJobWorker } from "./job-dispatch/index.js";
 import { redis } from "./redis.js";
-import { createReleaseNewVersionWorker } from "./releases/new-version/index.js";
-import { createReleaseVariableChangeWorker } from "./releases/variable-change/index.js";
-import { createResourceScanWorker } from "./resource-scan/index.js";
 import { workers } from "./workers/index.js";
 
 console.log("Registering instrumentation...");
 await register();
 
-const allWorkers = [
-  createResourceScanWorker(),
-  createDispatchExecutionJobWorker(),
-  createReleaseNewVersionWorker(),
-  createReleaseVariableChangeWorker(),
-  ...Object.values(workers),
-];
+const allWorkers = Object.values(workers);
 
 const shutdown = () => {
   logger.warn("Exiting...");
