+++ apps/event-worker/src/workers/job-dispatch/index.ts
@@ -1,0 +1,49 @@
+import { eq, takeFirstOrNull } from "@ctrlplane/db";
+import { db } from "@ctrlplane/db/client";
+import * as schema from "@ctrlplane/db/schema";
+import { createWorker } from "@ctrlplane/events";
+import { updateJob } from "@ctrlplane/job-dispatch";
+import { Channel } from "@ctrlplane/validators/events";
+import { JobAgentType, JobStatus } from "@ctrlplane/validators/jobs";
+
+import { redis } from "../../redis.js";
+import { dispatchGithubJob } from "./github.js";
+
+export const dispatchJobWorker = createWorker(
+  Channel.DispatchJob,
+  async (job) => {
+    const { jobId } = job.data;
+    const je = await db
+      .select()
+      .from(schema.job)
+      .innerJoin(schema.jobAgent, eq(schema.job.jobAgentId, schema.jobAgent.id))
+      .where(eq(schema.job.id, jobId))
+      .then(takeFirstOrNull);
+
+    if (je == null) {
+      job.log(`Job ${jobId} not found`);
+      return;
+    }
+
+    try {
+      job.log(
+        `Dispatching job ${je.job.id} --- ${je.job_agent.type}/${je.job_agent.name}`,
+      );
+      if (je.job_agent.type === String(JobAgentType.GithubApp)) {
+        job.log(`Dispatching to GitHub app`);
+        await dispatchGithubJob(je.job);
+      }
+    } catch (error: unknown) {
+      await updateJob(db, je.job.id, {
+        status: JobStatus.Failure,
+        message: (error as Error).message,
+      });
+    }
+  },
+  {
+    connection: redis,
+    removeOnComplete: { age: 1 * 60 * 60, count: 100 },
+    removeOnFail: { age: 12 * 60 * 60, count: 100 },
+    concurrency: 10,
+  },
+);
