+++ apps/event-worker/src/workers/resource-scan/aws/aws.ts
