+++ apps/event-worker/src/workers/resource-scan/google/client.ts
